Move stub Java services to `given`
